Rename merkle dir to commitment
